backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-multipart
httpx
python-dotenv
aiofiles
fpdf2
cloudinary
